pass sprite, keys and gamepad into player factory instead of game
